+++ src/pages/dataApprovalLevels/List.spec.tsx
@@ -1,13 +1,13 @@
-import schemaMock from '../../__mocks__/schema/programs.json'
+import schemaMock from '../../__mocks__/schema/dataApprovalLevel.json'
 import { SECTIONS_MAP } from '../../lib'
-import { testPrograms } from '../../testUtils/builders'
+import { testdataApprovalLevelList } from '../../testUtils/builders'
 import { generateDefaultListTests } from '../defaultListTests'
 import { Component } from './List'
 
-const section = SECTIONS_MAP.program
+const section = SECTIONS_MAP.dataApprovalLevel
 const mockSchema = schemaMock
 const ComponentToTest = Component
-const generateRandomElement = testPrograms
+const generateRandomElement = testdataApprovalLevelList
 const customData = {}
 
 generateDefaultListTests({