+++ src/pages/programsWip/Edit.tsx
@@ -38,8 +38,8 @@ const fieldFilters = [
     'categoryCombo[id,displayName]',
     'lastUpdated',
     'dataEntryForm[id,displayName,htmlCode]',
-    'programTrackedEntityAttributes',
     'programSections[id,displayName,description,access,sortOrder]',
+    'programTrackedEntityAttributes[id,displayName,valueType,renderType,allowFutureDate,mandatory,searchable,displayInList,trackedEntityAttribute[id,displayName]]',
 ] as const
 
 export type ProgramsFromFilters = PickWithFieldFilters<